Coverity fixes

+++ src/osgEarth/Memory.cpp
@@ -139,7 +139,7 @@ Memory::getProcessUsage()
         return (size_t)0L;      /* Can't read? */
     }
     fclose( fp );
-    return (size_t)rss * (size_t)sysconf( _SC_PAGESIZE);
+    return (unsigned) ((size_t)rss * (size_t)sysconf( _SC_PAGESIZE));
 
 #else
     /* AIX, BSD, Solaris, and Unknown OS ------------------------ */
